requirements.txt
alembic==1.12.1
click==8.1.7
colorama==0.4.6
Flask==2.2.5
flask-marshmallow==0.15.0
Flask-Migrate==4.0.7
flask-sqlalchemy==3.0.5
greenlet==3.1.1
importlib-metadata==6.7.0
importlib-resources==5.12.0
itsdangerous==2.1.2
jinja2==3.1.4
Mako==1.2.4
MarkupSafe==2.1.5
marshmallow==3.19.0
packaging==24.0
psycopg2-binary==2.9.9
python-dotenv==0.21.1
SQLAlchemy==2.0.36
typing-extensions==4.7.1
Werkzeug==2.2.3
zipp==3.15.0
